added stored procedures
a list of ingredients for a recipe given its name
gets client info when email and passwords match
 registers a consumed product by a client given the user email,
the date, meal time and the id of the consumed product
registers a consumed recipe by a client given the user email,
the date, meal time and the id of the consumed recipe
Registers the fisiological data given by the user in a specific date.
It acceps various null values. If a registry is given by the user in the
same date it overrides the values that are already stored
add new product